Refactored to split controller into reviews and categories
