--- controllers/app.controller.js
@@ -1,21 +1,0 @@
-const { fetchCategories, fetchReviewById } = require("../models/app.model.js");
-
-const getCategories = (req, res, next) => {
-  fetchCategories()
-    .then((categories) => {
-      res.status(200).send({ categories });
-    })
-    .catch(next);
-};
-
-const getReviewById = (req, res, next) => {
-  const id = req.params.review_id;
-
-  fetchReviewById(id)
-    .then((review) => {
-      res.status(200).send({ review });
-    })
-    .catch(next);
-};
-
-module.exports = { getCategories, getReviewById };
